services/modulo_inventarios/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
motor>=3.3.0
nats-py>=2.6.0
pydantic>=2.0.0
